test(thumbnail): migrate
