+++ core/src/components/thumbnail/test/basic/thumbnail.e2e.ts-snapshots/thumbnail-img-diff-ios-ltr-Mobile-Safari-linux.png
