+++ core/src/components/thumbnail/test/basic/thumbnail.e2e.ts-snapshots/thumbnail-ion-item-diff-ios-ltr-Mobile-Chrome-linux.png
