+++ core/src/components/thumbnail/test/basic/thumbnail.e2e.ts-snapshots/thumbnail-ion-item-diff-md-ltr-Mobile-Safari-linux.png
