+++ core/src/components/thumbnail/test/basic/thumbnail.e2e.ts-snapshots/thumbnail-ion-item-size-diff-ios-ltr-Mobile-Safari-linux.png
